# ARROW-6411: [Python][Parquet] Improve performance of DictEncoder::PutIndices