@wesm 2019-09-02
Here's a different processor (Xeon E-2176M), still compiling with clang 7.0

before

```
[ 25.00%] ··· parquet.ParquetWriteDictionaries.time_write_random_order                                                                                                                       ok
[ 25.00%] ··· ========= =========
               nunique
              --------- ---------
                 1000    1.65±0s
                100000   1.53±0s
              ========= =========

[ 50.00%] ··· parquet.ParquetWriteDictionaries.time_write_sequential                                                                                                                         ok
[ 50.00%] ··· ========= =========
               nunique
              --------- ---------
                 1000    1.31±0s
                100000   1.04±0s
              ========= =========
```

after

```
[ 25.00%] ··· parquet.ParquetWriteDictionaries.time_write_random_order                                                                                                                       ok
[ 25.00%] ··· ========= =========
               nunique
              --------- ---------
                 1000    891±0ms
                100000   1.16±0s
              ========= =========

[ 50.00%] ··· parquet.ParquetWriteDictionaries.time_write_sequential                                                                                                                         ok
[ 50.00%] ··· ========= =========
               nunique
              --------- ---------
                 1000    566±0ms
                100000   720±0ms
              ========= =========
```

The perf issue is also present with gcc 8.3, which net performs better on this machine

before

```
[ 25.00%] ··· parquet.ParquetWriteDictionaries.time_write_random_order                                                                                                                       ok
[ 25.00%] ··· ========= =========
               nunique           
              --------- ---------
                 1000    1.57±0s 
                100000   1.40±0s 
              ========= =========

[ 50.00%] ··· parquet.ParquetWriteDictionaries.time_write_sequential                                                                                                                         ok
[ 50.00%] ··· ========= =========
               nunique           
              --------- ---------
                 1000    1.28±0s 
                100000   955±0ms 
              ========= =========
```

after

```
[ 25.00%] ··· parquet.ParquetWriteDictionaries.time_write_random_order                              
                                                                                         ok
[ 25.00%] ··· ========= =========
               nunique
              --------- ---------
                 1000    695±0ms
                100000   988±0ms
              ========= =========

[ 50.00%] ··· parquet.ParquetWriteDictionaries.time_write_sequential                                                                                                                         ok
[ 50.00%] ··· ========= =========
               nunique
              --------- ---------
                 1000    418±0ms
                100000   564±0ms
              ========= =========
```