@wesm 2019-09-02
I'm going to add a C++ benchmark so this is easy to profile without having to go through Python