@wesm 2019-09-02
C++ benchmark results

before

```
----------------------------------------------------------------------------------------
Benchmark                                                 Time           CPU Iterations
----------------------------------------------------------------------------------------
BM_ArrowBinaryDict/EncodeDictDirectInt8/1048576     7334087 ns    7333876 ns         98   136.354M items/s
BM_ArrowBinaryDict/EncodeDictDirectInt16/1048576    7022430 ns    7022412 ns        100   142.401M items/s
BM_ArrowBinaryDict/EncodeDictDirectInt32/1048576    7061033 ns    7060870 ns         99   141.626M items/s
BM_ArrowBinaryDict/EncodeDictDirectInt64/1048576    7084581 ns    7084398 ns         97   141.155M items/s
```

after

```
----------------------------------------------------------------------------------------
Benchmark                                                 Time           CPU Iterations
----------------------------------------------------------------------------------------
BM_ArrowBinaryDict/EncodeDictDirectInt8/1048576     4387151 ns    4387175 ns        156   227.937M items/s
BM_ArrowBinaryDict/EncodeDictDirectInt16/1048576    4446167 ns    4446074 ns        159   224.918M items/s
BM_ArrowBinaryDict/EncodeDictDirectInt32/1048576    4501028 ns    4500934 ns        156   222.176M items/s
BM_ArrowBinaryDict/EncodeDictDirectInt64/1048576    4635792 ns    4635728 ns        150   215.716M items/s
```