@wesm 2019-09-03
+1. I welcome further analysis of this issue in subsequent JIRA issues, etc.